refactored code

+++ app/src/main/java/org/example/entity/request/ConvertRequest.java
@@ -4,6 +4,8 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 @Schema(description = "Details about currency conversion request")
 public record ConvertRequest(
         @Schema(description = "From currency code", example = "USD", type = "string")
@@ -15,4 +17,4 @@ public record ConvertRequest(
         @Schema(description = "Amount to convert", example = "10.4", type = "double")
         @Positive(message = "amount must be positive")
         @NotNull(message = "amount is required")
-        Double amount) { }
+        BigDecimal amount) { }
